Move services to separate package, clean up
