+++ go.mod
@@ -3,9 +3,6 @@ module github.com/nats-io/nats.go
 go 1.16
 
 require (
-	github.com/golang/protobuf v1.5.0
-	github.com/nats-io/nats-server/v2 v2.9.3 // indirect
 	github.com/nats-io/nkeys v0.3.0
 	github.com/nats-io/nuid v1.0.1
-	google.golang.org/protobuf v1.28.1
 )
